fix recaptcha load components
